Update delete default records and update migration
