--- cmd/go_build_github_com_PurpleSchoolPractice_metiing_pro_golang
@@ -1,1 +1,0 @@
-ELF
