+++ cmd/root.go
@@ -1,11 +1,10 @@
-package main
+package cmd
 
 import (
 	"fmt"
 	"os"
 
 	"github.com/PurpleSchoolPractice/metiing-pro-golang/configs"
-	"github.com/PurpleSchoolPractice/metiing-pro-golang/migrations"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -21,21 +20,12 @@ var appConfig Cfg
 
 // rootCmd представляет собой базовую команду при вызове без подкоманд
 var rootCmd = &cobra.Command{
-	Use:   "meeting-pro",
+	Use:   "meeting",
 	Short: "Meeting Pro - приложение для управления встречами",
 	Long:  `Meeting Pro - приложение для управления встречами и расписаниями встреч.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		// Загружаем конфигурацию с Viper
 		initConfig()
-	},
-}
-var downgrade = &cobra.Command{
-	Use:   "deldefrec",
-	Short: "Delete default records from the database",
-	Long:  `This command deletes default records from users, secrets, events, and event_participants tables.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("Running...")
-		migrations.Migrate()
 	},
 }
 
@@ -63,7 +53,7 @@ func init() {
 	rootCmd.PersistentFlags().String("log-level", "info", "уровень логирования (debug, info, warn, error)")
 
 	viper.AutomaticEnv()
-	rootCmd.AddCommand(downgrade)
+
 }
 
 func initConfig() {
